Save state (bin/sql)
Worked some more on integration of database functions, as well as ad hoc testing.

Next steps are to continue implementing database integration: working on the SQL if it's still not finished, otherwise working on the binary file writer/reader for local saves.

I managed to get the binary file saving and loading done, next steps are to add in a check against the database to see which save method is the most up to date and then use that one.